requirements.txt
streamlit>=1.28.0
beautifulsoup4>=4.12.0
requests>=2.31.0
google-api-python-client>=2.100.0
sentence-transformers>=2.2.0
openai>=1.0.0
python-dateutil>=2.8.0
